Create sp-weights crate to store weight primitives
