+++ Cargo.toml
@@ -211,6 +211,7 @@ members = [
 	"primitives/version",
 	"primitives/version/proc-macro",
 	"primitives/wasm-interface",
+	"primitives/weights",
 	"test-utils/client",
 	"test-utils/derive",
 	"test-utils/runtime",
