+++ frame/transaction-payment/asset-tx-payment/src/lib.rs
@@ -39,7 +39,7 @@ use sp_std::prelude::*;
 
 use codec::{Decode, Encode};
 use frame_support::{
-	dispatch::DispatchResult,
+	dispatch::{DispatchInfo, DispatchResult, PostDispatchInfo},
 	traits::{
 		tokens::{
 			fungibles::{Balanced, CreditOf, Inspect},
@@ -47,7 +47,6 @@ use frame_support::{
 		},
 		IsType,
 	},
-	weights::{DispatchInfo, PostDispatchInfo},
 	DefaultNoBound,
 };
 use pallet_transaction_payment::OnChargeTransaction;
