+++ frame/transaction-payment/src/lib.rs
@@ -63,11 +63,11 @@ use sp_runtime::{
 use sp_std::prelude::*;
 
 use frame_support::{
-	dispatch::DispatchResult,
-	traits::{EstimateCallFee, Get},
-	weights::{
-		DispatchClass, DispatchInfo, GetDispatchInfo, Pays, PostDispatchInfo, Weight, WeightToFee,
+	dispatch::{
+		DispatchClass, DispatchInfo, DispatchResult, GetDispatchInfo, Pays, PostDispatchInfo,
 	},
+	traits::{EstimateCallFee, Get},
+	weights::{Weight, WeightToFee},
 };
 
 mod payment;
@@ -822,12 +822,11 @@ mod tests {
 	};
 
 	use frame_support::{
-		assert_noop, assert_ok, parameter_types,
+		assert_noop, assert_ok,
+		dispatch::{DispatchClass, DispatchInfo, GetDispatchInfo, PostDispatchInfo},
+		parameter_types,
 		traits::{ConstU32, ConstU64, Currency, GenesisBuild, Imbalance, OnUnbalanced},
-		weights::{
-			DispatchClass, DispatchInfo, GetDispatchInfo, PostDispatchInfo, Weight,
-			WeightToFee as WeightToFeeT,
-		},
+		weights::{Weight, WeightToFee as WeightToFeeT},
 	};
 	use frame_system as system;
 	use pallet_balances::Call as BalancesCall;
