+++ utils/frame/benchmarking-cli/src/overhead/weights.hbs
@@ -29,10 +29,8 @@
 //   {{arg}}
 {{/each}}
 
-use frame_support::{
-	parameter_types,
-	weights::{constants::WEIGHT_PER_NANOS, Weight},
-};
+use sp_weights::{constants::WEIGHT_PER_NANOS, Weight};
+use sp_core::parameter_types;
 
 parameter_types! {
 	{{#if (eq short_name "block")}}
@@ -58,7 +56,7 @@ parameter_types! {
 #[cfg(test)]
 mod test_weights {
 	use super::*;
-	use frame_support::weights::constants;
+	use sp_weights::constants;
 
 	/// Checks that the weight exists and is sane.
 	// NOTE: If this test fails but you are sure that the generated values are fine,
